[EDIFIKANA] #225 Added more functions to rbacService, added authorization checks to controllers.

+++ edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/core/controller/PropertyController.kt
@@ -6,8 +6,12 @@ import com.cramsan.edifikana.lib.model.OrganizationId
 import com.cramsan.edifikana.lib.model.PropertyId
 import com.cramsan.edifikana.lib.model.network.CreatePropertyNetworkRequest
 import com.cramsan.edifikana.lib.model.network.UpdatePropertyNetworkRequest
+import com.cramsan.edifikana.lib.utils.ClientRequestExceptions
+import com.cramsan.edifikana.server.core.controller.authentication.ClientContext
 import com.cramsan.edifikana.server.core.controller.authentication.ContextRetriever
 import com.cramsan.edifikana.server.core.service.PropertyService
+import com.cramsan.edifikana.server.core.service.authorization.RoleBasedAccessControlService
+import com.cramsan.edifikana.server.core.service.models.UserRole
 import com.cramsan.framework.annotations.NetworkModel
 import com.cramsan.framework.core.ktor.HttpResponse
 import io.ktor.http.HttpStatusCode
@@ -26,6 +30,7 @@ import io.ktor.server.routing.route
  */
 class PropertyController(
     private val propertyService: PropertyService,
+    private val rbacService: RoleBasedAccessControlService,
     private val contextRetriever: ContextRetriever,
 ) : Controller {
 
@@ -39,6 +44,11 @@ class PropertyController(
             "createProperty",
             contextRetriever
         ) { context ->
+            if (!rbacService.hasRoleOrHigher(context, UserRole.OWNER)) {
+                throw ClientRequestExceptions.UnauthorizedException(
+                    "You do not have permissions to create new properties."
+                )
+            }
             val createPropertyRequest = call.receive<CreatePropertyNetworkRequest>()
 
         val newProperty = propertyService.createProperty(
@@ -64,6 +74,7 @@ class PropertyController(
         contextRetriever
     ) { context ->
         val propertyId = requireNotNull(call.parameters[PROPERTY_ID])
+        checkAuthorization(context, PropertyId(propertyId), UserRole.MANAGER)
 
         val property = propertyService.getProperty(
             PropertyId(propertyId),
@@ -83,6 +94,7 @@ class PropertyController(
 
     /**
      * Handles the retrieval of a list of properties. The [call] parameter is the request context.
+     * TODO: ADD A CHECK THAT THE REQUESTER IS REQUESTING ALL PROPERTIES FROM THEIR ORG
      */
     @OptIn(NetworkModel::class)
     suspend fun getProperties(call: ApplicationCall) = call.handleCall(
@@ -90,6 +102,13 @@ class PropertyController(
         "getProperties",
         contextRetriever,
     ) { context ->
+        if (!rbacService.hasRole(context, UserRole.OWNER)
+        ) {
+            throw ClientRequestExceptions.UnauthorizedException(
+                "You do not have permissions to see all properties in your organization."
+            )
+        }
+
         val properties = propertyService.getProperties(
             userId = context.userId,
         ).map { it.toPropertyNetworkResponse() }
@@ -110,6 +129,7 @@ class PropertyController(
         contextRetriever
     ) { context ->
         val propertyId = requireNotNull(call.parameters[PROPERTY_ID])
+        checkAuthorization(context, PropertyId(propertyId), UserRole.OWNER)
 
         val updatePropertyRequest = call.receive<UpdatePropertyNetworkRequest>()
 
@@ -133,6 +153,7 @@ class PropertyController(
         contextRetriever
     ) { context ->
         val propertyId = requireNotNull(call.parameters[PROPERTY_ID])
+        checkAuthorization(context, PropertyId(propertyId), UserRole.OWNER)
 
         val success = propertyService.deleteProperty(
             PropertyId(propertyId),
@@ -148,6 +169,21 @@ class PropertyController(
             status = statusCode,
             body = null,
         )
+    }
+
+    /**
+     * Checks if the user in the [context] has at least the [requiredRole] for the [targetProperty].
+     */
+    private suspend fun checkAuthorization(
+        context: ClientContext.AuthenticatedClientContext,
+        targetProperty: PropertyId,
+        requiredRole: UserRole,
+    ) {
+        if (!rbacService.hasRoleOrHigher(context, targetProperty, requiredRole)) {
+            throw ClientRequestExceptions.UnauthorizedException(
+                "You do not have permissions to edit/delete a property."
+            )
+        }
     }
 
     /**
