+++ edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/core/controller/StaffController.kt
@@ -6,8 +6,12 @@ import com.cramsan.edifikana.lib.model.PropertyId
 import com.cramsan.edifikana.lib.model.StaffId
 import com.cramsan.edifikana.lib.model.network.CreateStaffNetworkRequest
 import com.cramsan.edifikana.lib.model.network.UpdateStaffNetworkRequest
+import com.cramsan.edifikana.lib.utils.ClientRequestExceptions
+import com.cramsan.edifikana.server.core.controller.authentication.ClientContext
 import com.cramsan.edifikana.server.core.controller.authentication.ContextRetriever
 import com.cramsan.edifikana.server.core.service.StaffService
+import com.cramsan.edifikana.server.core.service.authorization.RoleBasedAccessControlService
+import com.cramsan.edifikana.server.core.service.models.UserRole
 import com.cramsan.framework.annotations.NetworkModel
 import com.cramsan.framework.core.ktor.HttpResponse
 import io.ktor.http.HttpStatusCode
@@ -26,6 +30,7 @@ import io.ktor.server.routing.route
  */
 class StaffController(
     private val staffService: StaffService,
+    private val rbacService: RoleBasedAccessControlService,
     private val contextRetriever: ContextRetriever,
 ) : Controller {
 
@@ -38,6 +43,12 @@ class StaffController(
         "createStaff",
         contextRetriever,
     ) { context ->
+        // Only OWNERS can create new staff
+        if (!rbacService.hasRole(context, UserRole.OWNER)) {
+            throw ClientRequestExceptions.UnauthorizedException(
+                "You do not have permissions to create new staff."
+            )
+        }
         val createStaffRequest = call.receive<CreateStaffNetworkRequest>()
 
         val newStaff = staffService.createStaff(
@@ -64,6 +75,7 @@ class StaffController(
         contextRetriever,
     ) { context ->
         val staffId = requireNotNull(call.parameters[STAFF_ID])
+        checkAuthorization(context, StaffId(staffId), UserRole.MANAGER)
 
         val staff = staffService.getStaff(
             StaffId(staffId),
@@ -83,6 +95,7 @@ class StaffController(
 
     /**
      * Handles the retrieval of all staff. The [call] parameter is the request context.
+     * TODO: ADD A CHECK THAT THE REQUESTER IS REQUESTING ALL STAFF FROM THEIR ORG
      */
     @OptIn(NetworkModel::class)
     suspend fun getStaffs(call: ApplicationCall) = call.handleCall(
@@ -90,8 +103,14 @@ class StaffController(
         "getStaffs",
         contextRetriever,
     ) { context ->
-        val authenticatedClientContext = requireAuthenticatedClientContext(context)
-        val staffs = staffService.getStaffs(authenticatedClientContext).map { it.toStaffNetworkResponse() }
+        if (!rbacService.hasRoleOrHigher(context, UserRole.MANAGER)) {
+            throw ClientRequestExceptions.UnauthorizedException(
+                "You do not have permission to see all Staff for your organization."
+            )
+        }
+
+        val staffs = staffService.getStaffs(context).map { it.toStaffNetworkResponse() }
+
 
         HttpResponse(
             status = HttpStatusCode.OK,
@@ -109,6 +128,7 @@ class StaffController(
         contextRetriever,
     ) { context ->
         val staffId = requireNotNull(call.parameters[STAFF_ID])
+        checkAuthorization(context, StaffId(staffId), UserRole.MANAGER)
 
         val updateStaffRequest = call.receive<UpdateStaffNetworkRequest>()
 
@@ -135,6 +155,7 @@ class StaffController(
         contextRetriever
     ) { context ->
         val staffId = requireNotNull(call.parameters[STAFF_ID])
+        checkAuthorization(context, StaffId(staffId), UserRole.MANAGER)
 
         val success = staffService.deleteStaff(
             StaffId(staffId),
@@ -151,6 +172,22 @@ class StaffController(
             body = null,
         )
     }
+
+    /**
+     * Checks if the authenticated user has the required role or higher to perform actions on the target staff.
+     */
+    private suspend fun checkAuthorization(
+        context: ClientContext.AuthenticatedClientContext,
+        targetStaff: StaffId,
+        requiredRole: UserRole
+    ) {
+        if (!rbacService.hasRoleOrHigher(context, targetStaff, requiredRole)) {
+            throw ClientRequestExceptions.UnauthorizedException(
+                "You do not have permissions to edit/delete staff."
+            )
+        }
+    }
+
 
     /**
      * Registers the routes for the staff controller. The [route] parameter is the root path for the controller.
